+++ sensor_data.csv.dvc
@@ -1,5 +1,5 @@
 outs:
-- md5: cf3fb143538b41ebd732c29552a42457
-  size: 211721
+- md5: 4d60f91ae3384e31f4acf151c43fa149
+  size: 190029
   hash: md5
   path: sensor_data.csv
